requirements.txt
pyyaml
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
kaggle
kagglehub
ipywidgets
Pillow
nvidia-ml-py